Changed background, removed nav, and added add task back in

+++ contracts/TodoList.sol
@@ -1,47 +1,39 @@
 pragma solidity ^0.5.0;
 
 contract TodoList {
-  uint public taskCount = 0; /// everyone starts off with 6 tasks
+    uint256 public taskCount = 0; /// everyone starts off with 6 tasks
 
-  struct Task {
-    uint id;
-    string content;
-    bool completed;
-  }
+    struct Task {
+        uint256 id;
+        string content;
+        bool completed;
+    }
 
-  mapping(uint => Task) public tasks;
+    mapping(uint256 => Task) public tasks;
 
-  event TaskCreated(
-    uint id,
-    string content,
-    bool completed
-  );
+    event TaskCreated(uint256 id, string content, bool completed);
 
-  event TaskCompleted(
-    uint id,
-    bool completed
-  );
+    event TaskCompleted(uint256 id, bool completed);
 
-  constructor() public {
-    createTask("Compost leftovers!");
-    createTask("Reduce Reuse Recycle");
-    createTask("Limit shower time");
-    createTask("Turn unnecessary lights off");
-    createTask("Washed dishes by hand");
-    createTask("Carpooled/Walked/Biked/Used Public Transport");
-  }
+    constructor() public {
+        createTask("Compost leftovers!");
+        createTask("Reduce Reuse Recycle");
+        createTask("Limit shower time");
+        createTask("Turn unnecessary lights off");
+        createTask("Washed dishes by hand");
+        createTask("Carpooled/Walked/Biked/Used Public Transport");
+    }
 
-  function createTask(string memory _content) public {
-    taskCount ++;
-    tasks[taskCount] = Task(taskCount, _content, false);
-    emit TaskCreated(taskCount, _content, false);
-  }
+    function createTask(string memory _content) public {
+        taskCount++;
+        tasks[taskCount] = Task(taskCount, _content, false);
+        emit TaskCreated(taskCount, _content, false);
+    }
 
-  function toggleCompleted(uint _id) public {
-    Task memory _task = tasks[_id];
-    _task.completed = !_task.completed;
-    tasks[_id] = _task;
-    emit TaskCompleted(_id, _task.completed);
-  }
-
+    function toggleCompleted(uint256 _id) public {
+        Task memory _task = tasks[_id];
+        _task.completed = !_task.completed;
+        tasks[_id] = _task;
+        emit TaskCompleted(_id, _task.completed);
+    }
 }
